Nix Flake for dev and release tooling
The change turns the repo into a Nix flake (flake.nix + flake.lock).
Dependencies are now defined as flake inputs.

This doesn't influence runtime behaviour of the platform and we still
use versions.json/versions.nix for dependencies but versions.json is now
generated by the flake with the same content as before.

To modularize and provide convenient abstractions for the flake file,
especially when dealing with support for multiple systems/architectures
we use [flake-parts](https://flake.parts/).

[devenv](devenv.sh) is used to define the development shell which
replaces shell.nix and provides unified access to various helper
scripts. The shell works for Macs (release tooling is usually run from
there) and on Linux (typically developer VMs).

As the top-level dir became too crowded, many scripts and files
from the top-level which are related to dependency management and the
release process moved to the `release` subdir. Their location isn't
important anymore as they are meant to be called through scripts
generated by devenv now. `./update-nixpkgs.py` is now invoked as
`update_nixpkgs`, for example.  See the comment at the top of
`flake.nix` for more info.

./dev-setup still exists as script for our "muscle memory" but is only a
facade now, just calling `dev_setup` running in a dev shell.

During the the top-level directory cleanup, fc/ceph/__init__.py which
serves no purpose was removed, too. This causes isort (for Python),
which is part of the pre-commit hook, to sort imports differrently
as fc.ceph is now recognized as external dependency. This explains
the changes in pkgs/fc/ceph/fc/ceph.

PL-131814